Reorganize marginal cost scripts in utils
