+++ tests/test_bulk_tx_mc.py
@@ -9,7 +9,7 @@ from typing import cast
 import polars as pl
 import pytest
 
-from utils.pre.marginal_costs.bulk_tx_nyiso import (
+from utils.data_prep.marginal_costs.bulk_tx_nyiso import (
     DEFAULT_SCR_WINTER_MONTHS,
     aggregate_paying_locality_hourly_signals_from_constraint_groups,
     allocate_bulk_tx_to_hours,
