+++ tests/test_generate_utility_supply_mc.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import polars as pl
 import pytest
 
-from utils.pre.marginal_costs.supply_capacity_nyiso import (
+from utils.data_prep.marginal_costs.supply_capacity_nyiso import (
     allocate_icap_to_hours,
     build_locality_load_profiles,
     compute_components,
