+++ utils/data_prep/marginal_costs/bulk_tx_isone.py
@@ -37,7 +37,7 @@ from typing import cast
 
 import polars as pl
 
-from utils.pre.marginal_costs.supply_utils import (
+from utils.data_prep.marginal_costs.supply_utils import (
     allocate_annual_exceedance_to_hours,
     build_cairo_8760_timestamps,
 )
