+++ utils/data_prep/marginal_costs/supply_capacity_isone.py
@@ -31,7 +31,7 @@ from datetime import date
 
 import polars as pl
 
-from utils.pre.marginal_costs.supply_utils import (
+from utils.data_prep.marginal_costs.supply_utils import (
     DEFAULT_ISONE_FCA_S3_PATH,
     DEFAULT_ISONE_ZONE_LOADS_S3_BASE,
     ISONE_CAPACITY_ZONE_FALLBACK,
